Add files via upload
nltk_summarizer: This function uses NLTK, a natural language processing library, to generate a summary of text. It calculates word frequencies, normalizes them, and assigns scores to sentences based on their word frequencies. The top 7 sentences with the highest scores are selected as the summary.

text_summarizer: This function utilizes Spacy, a natural language processing library, to generate a summary of text. It calculates word frequencies, normalizes them, and assigns scores to sentences based on their word frequencies. The top 7 sentences with the highest scores are selected as the summary.